Added lines to autoscrp to deal with dates on republish.
If a date is already provided in the file `date.txt` then this is
used as the date for the post. (The file will be set up by
`archive.sh`.) If `date.txt` does not exist, it will pull the date
from the string provided as a general argument.

This is a step towards allowing republishing posts after minor edits
without having to make the post new, or bringing it up the blog.