Changed way dates are handled so that more standard is accepted.
Previously, the date in `date.txt` had to be formatted so as
to fit with my current conventions for folder and file names
for archived materials.

I've now changed this so that the date in `date.txt` is now
a more conventional ISO format. The script `archive.sh` has to do
a bit more work to make it ready for formatting but `autoscrp.sh`
now just takes it directly from the file.