Added basic image handling to `autoscrp.sh`.
When publishing (or re-publishing) a post to Scriptogram,
a map of available image keys and the addresses is built
from a csv file and the post file is parsed for
indications of local image references (prefixed by `/!/`)
in Markdown image placements.

These local image references are used as a key to the map
and replaced with the address values so the the published
version on Scriptogram can show the hosted versions of
each image.

This is a first step towards automating handling of
images.