Ensured file references in `test-post-photo.sh` are more stable.
Did the same for `refresh-access-token.sh`.

Previously references to other files were relative. This meant
if the script was run from another directory, the script
could not find the files it needed.

I've taken the same approach here as I have with other scripts
by finding the directory the actual script being executed sits
in and assigning this to a `$DIR` variable.

I also moved the writing to `imagemap.csv` to within an `if`
condition so that an error gets produced rather than writing
blank lines to that file.