Did a major rearrange of the files and folders.
This was to separate content from code and from userdata.

Content should be separate from the code so that the model can be
reused for several different projects.

Userdata should be separate from the code so that the model
can be used by more than one user without revealing
account details.

Code is the main focus of versioning in this repository so that
the model is captured.

Three folders `content`, `userdata` and `code` have been created
to this end.